+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### Fixed
 - Fixed a bug with campaign and mailing list charts not appearing correctly in reports tab on edit pages
+- Fixed a bug in the first sent date and time in campaign reports 
 - Fixed a bug with contact activity in mailing list reports
 
 > {tip} This version adds multi-site functionality to campaign types, mailing list types, segments and sendouts. All existing elements will be assigned to the primary site by default.